// frontend/src_backup/types.tsx
export interface ApiResponse<T> {
  success: boolean
  message: string
  data: T
  error?: string
}